rm lazy and _subset_anndata
