+++ lamindb/dev/__init__.py
@@ -6,12 +6,9 @@
    BaseORM
    QuerySet
    datasets
-   LazyDataFrame
 """
 
 from lnschema_core._queryset import QuerySet
 from lnschema_core.models import BaseORM
-
-from lamindb.dev.storage.object._lazy_field import Lazy as LazyDataFrame
 
 from . import datasets  # noqa
