+++ lamindb/dev/storage/__init__.py
@@ -25,5 +25,4 @@ from lamindb_setup.dev.upath import infer_filesystem as _infer_filesystem
 from ._file import delete_storage, load_to_memory, store_object
 from ._h5ad import h5ad_to_anndata
 from ._images import store_png
-from ._subset import subset
 from ._zarr import read_adata_zarr, write_adata_zarr
